moves displaybox code from components to ui
